add graceful shutdown for agent
