+++ internal/common/common.go
@@ -6,6 +6,9 @@ import (
 	"encoding/hex"
 	"fmt"
 	"log"
+	"os"
+	"os/signal"
+	"syscall"
 )
 
 const (
@@ -19,6 +22,37 @@ type Metrics struct {
 	Delta *int64   `json:"delta,omitempty"` // значение метрики в случае передачи counter
 	Value *float64 `json:"value,omitempty"` // значение метрики в случае передачи gauge
 	Hash  string   `json:"hash,omitempty"`  // значение хеш-функции
+}
+
+func AwaitInterruptSignal(exitChan chan<- int) {
+	log.Println("Awaiting interrupt signal")
+
+	signalChanel := make(chan os.Signal, 1)
+	signal.Notify(signalChanel,
+		syscall.SIGINT,
+		syscall.SIGTERM,
+		syscall.SIGQUIT)
+
+	go func() {
+		s := <-signalChanel
+		switch s {
+		case syscall.SIGINT:
+			log.Printf("%s SIGINT signal triggered.", s)
+			exitChan <- 1
+
+		case syscall.SIGTERM:
+			log.Printf("%s SIGTERM signal triggered.", s)
+			exitChan <- 2
+
+		case syscall.SIGQUIT:
+			log.Printf("%s SIGQUIT signal triggered.", s)
+			exitChan <- 3
+
+		default:
+			log.Printf("%s UNKNOWN signal triggered.", s)
+			exitChan <- 1
+		}
+	}()
 }
 
 func (m Metrics) computeHash(key string) ([]byte, error) {
